requirements.txt
pyyaml
# mustache templating
chevron

# for yaml parsing
# pyyaml destroys string quotes
ruamel.yaml